Use `svn info` to avoid a round-trip to the SVN server.
Note that because the `--xml` output of `svn info` is not stable, we (finally) are pushed to using a proper XML parser to read the output. The stdlib SAX parser is available to all python versions we support, and is easy enough to work with. The change is big enough to call this version 3.0.

Credit for the `svn info` idea goes to mgood (Matt Good). This closes issue #6.